push

+++ app/Http/Controllers/FakultasController.php
@@ -30,6 +30,11 @@ class FakultasController extends Controller
      */
     public function store(Request $request)
     {
+        // cek apakah user memiliki izin untuk membuat fakultas
+        if($request->user()->cannot('create', Fakultas::class)){
+            abort(403);
+        }
+
         // validasi input
         $input = $request->validate([
             'nama' => 'required|unique:fakultas',
@@ -74,6 +79,11 @@ class FakultasController extends Controller
     public function update(Request $request,   $fakultas)
     {
         $fakultas = Fakultas::findOrFail($fakultas);
+
+        if ($request->user()->cannot('update', $fakultas)){
+            abort(403);
+        }
+        
         // validasi input
         $input = $request->validate([
             'nama' => 'required',
@@ -90,10 +100,14 @@ class FakultasController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy($fakultas)
+    public function destroy(Request $request, $fakultas)
     {
         $fakultas = Fakultas::findOrFail($fakultas);
         // dd($fakultas);
+
+         if ($request->user()->cannot('delete', $fakultas)){
+            abort(403);
+        }
 
         // hapus data fakultas
         $fakultas->delete();
